refactor: move user-related files to /users folder
Relocate user-related components, state, hooks, and effects into the /users
folder to improve organization and adhere to a colocation-based file structure.
