+++ components/layouts/layoutHeader/index.tsx
@@ -14,8 +14,8 @@ import { Logo } from './logo';
 import { NavigationButton } from './navigationButton';
 import { atomNavigationOpen } from '@states/layouts';
 
-const UserSessionGroupEffects = dynamic(() =>
-  import('@effects/users').then((mod) => mod.UserSessionGroupEffects),
+const UserSessionGroupEffect = dynamic(() =>
+  import('@users/userSessionGroupEffect').then((mod) => mod.UserSessionGroupEffect),
 );
 
 type Props = Partial<Pick<Types, 'children'>> & Pick<Types, 'layoutType'>;
@@ -35,28 +35,27 @@ export const LayoutHeader = ({ children, layoutType }: Props) => {
           layoutHome && 'z-50 bg-slate-50',
           homeSidebarClose && 'bg-opacity-60 backdrop-blur-lg',
           layoutApp && 'bg-transparent',
-        )}>
+        )}
+      >
         <div
           className={classNames(
             'flex max-h-[4rem] min-h-[4rem] flex-row items-center justify-between ml:mb-2',
             layoutHome && 'mx-auto max-w-7xl',
-          )}>
+          )}
+        >
           <LeftSideFragment>
             <div
               className={classNames(
                 'flex flex-row items-center justify-between pl-3 md:w-full',
                 layoutApp && 'md:max-w-[16rem]',
                 layoutHome && 'md:max-w-[12rem]',
-              )}>
-              <NavigationButtonFragment>
-                {layoutApp && <NavigationButton />}
-              </NavigationButtonFragment>
+              )}
+            >
+              <NavigationButtonFragment>{layoutApp && <NavigationButton />}</NavigationButtonFragment>
               <LogoFragment>
                 <div
-                  className={classNames(
-                    'flex w-full flex-row justify-start',
-                    layoutApp && 'max-md:hidden',
-                  )}>
+                  className={classNames('flex w-full flex-row justify-start', layoutApp && 'max-md:hidden')}
+                >
                   <Logo />
                 </div>
               </LogoFragment>
@@ -67,23 +66,19 @@ export const LayoutHeader = ({ children, layoutType }: Props) => {
               className={classNames(
                 layoutApp && 'flex flex-1 flex-row items-center justify-end pl-2 ml:mr-3',
                 layoutHome && 'hidden ml:mr-8 ml:flex',
-              )}>
+              )}
+            >
               {children}
             </div>
             <NavigationButtonFragment>
-              <div
-                className={classNames(
-                  'ml:hidden ml:pr-0',
-                  layoutHome && 'pr-6',
-                  layoutApp && 'pr-3',
-                )}>
+              <div className={classNames('ml:hidden ml:pr-0', layoutHome && 'pr-6', layoutApp && 'pr-3')}>
                 {layoutHome && <NavigationButton />}
               </div>
             </NavigationButtonFragment>
           </RightNavigationFragment>
         </div>
       </div>
-      <UserSessionGroupEffects />
+      <UserSessionGroupEffect />
     </LayoutHeaderFragment>
   );
 };
