+++ components/users/user/user.states.ts
@@ -1,0 +1,5 @@
+import { atom } from 'recoil';
+
+export const atomUserSession = atom({
+  key: 'atomUserSession',
+});
