+++ components/users/user/userDropdown/index.tsx
@@ -5,20 +5,30 @@ import { nextImageLoader } from '@stateLogics/utils';
 import { signOut, useSession } from 'next-auth/react';
 import Image from 'next/image';
 import { Fragment } from 'react';
-import { Dropdown } from './dropdown';
-import { MenuItem } from './dropdown/menuItem';
+import { Dropdown } from '@dropdowns/v2/dropdown';
+import { MenuItem } from '@dropdowns/v2/dropdown/menuItem';
 
 export const UserDropdown = () => {
   const { data: session } = useSession();
   const userImage = session?.user.image;
+  const dropdownOptions = {
+    hasDivider: false,
+    padding: 'p-0',
+    hoverRing: 'hover:ring-4 hover:ring-slate-200',
+  };
+  const menuItemSettingOptions = {
+    path: ICON_SETTINGS,
+    tooltip: 'Settings',
+    isDisabled: true,
+  };
+  const menuItemSignOutOptions = {
+    path: ICON_LOGOUT,
+    tooltip: 'Sign out',
+  };
 
   return (
     <Dropdown
-      options={{
-        hasDivider: false,
-        padding: 'p-0',
-        hoverRing: 'hover:ring-4 hover:ring-slate-200',
-      }}
+      options={dropdownOptions}
       menuButtonContent={
         <Fragment>
           <span className='sr-only'>Open user menu</span>
@@ -32,26 +42,18 @@ export const UserDropdown = () => {
             alt='User avatar'
           />
         </Fragment>
-      }>
+      }
+    >
       <ActiveDropdownMenuItemEffect menuItemId={null} />
       {/* give menuItemId any ID: string to activate the keyboard navigation */}
       <div className='py-1'>
-        <MenuItem
-          options={{
-            path: ICON_SETTINGS,
-            tooltip: 'Settings',
-            isDisabled: true,
-          }}>
-          Settings
-        </MenuItem>
+        <MenuItem options={menuItemSettingOptions}>Settings</MenuItem>
       </div>
       <div className='py-1'>
         <MenuItem
-          options={{
-            path: ICON_LOGOUT,
-            tooltip: 'Sign out',
-          }}
-          onClick={() => signOut({ callbackUrl: process.env.NEXT_PUBLIC_HOST })}>
+          options={menuItemSignOutOptions}
+          onClick={() => signOut({ callbackUrl: process.env.NEXT_PUBLIC_HOST })}
+       >
           Sign out
         </MenuItem>
       </div>
