+++ components/users/userSessionGroupEffect/index.tsx
@@ -1,0 +1,11 @@
+import { UserSessionEffect } from './userSessionEffect';
+import { UserSessionResetEffect } from './userSessionResetEffect';
+
+export const UserSessionGroupEffect = () => {
+  return (
+    <>
+      <UserSessionEffect />
+      <UserSessionResetEffect />
+    </>
+  );
+};
