+++ tsconfig.json
@@ -11,6 +11,7 @@
       "@data/*": ["lib/data/*"],
       "@stateLogics/*": ["lib/stateLogics/*"],
       "@lib/*": ["lib/*"],
+      "@users/*": ["components/users/*"],
       "@layouts/*": ["components/layouts/*"],
       "@buttons/*": ["components/ui/buttons/*"],
       "@dropdowns/*": ["components/ui/dropdowns/*"],
